+++ deepparse/tools.py
@@ -27,6 +27,7 @@ def download_from_url(model: str, saving_dir: str, extension: str):
     r = requests.get(url)
 
     os.makedirs(saving_dir, exist_ok=True)
+
     open(os.path.join(saving_dir, f"{model}.{extension}"), "wb").write(r.content)