+++ deepparse/vectorizer/fasttext_vectorizer.py
@@ -30,5 +30,5 @@ class FastTextVectorizer(Vectorizer):
         Return:
             A list of word vector.
         """
-
+        address = address.replace(",", "")  # see issue 56 https://github.com/GRAAL-Research/deepparse/issues/56
         return [self.embeddings_model(word) for word in address.split()]