+++ CHANGELOG.md
@@ -12,3 +12,6 @@ Fixed authors in setup
 
 ## dev
 Added "contributing to"
+Added fix for comma problem (#56)
+Added content in Address Parser doc for tags definition
+Fixed Pylint bug with PyTorch 1.6